up to revision 28

+++ aCis_gameserver/java/net/sf/l2j/gameserver/handler/itemhandlers/PetFood.java
@@ -1,32 +1,16 @@
-/*
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later
- * version.
- * 
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- * 
- * You should have received a copy of the GNU General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
- */
 package net.sf.l2j.gameserver.handler.itemhandlers;
 
 import net.sf.l2j.Config;
-import net.sf.l2j.gameserver.datatables.PetDataTable;
-import net.sf.l2j.gameserver.datatables.SkillTable;
+import net.sf.l2j.gameserver.data.SkillTable;
 import net.sf.l2j.gameserver.handler.IItemHandler;
 import net.sf.l2j.gameserver.model.L2Skill;
-import net.sf.l2j.gameserver.model.actor.L2Playable;
-import net.sf.l2j.gameserver.model.actor.instance.L2PcInstance;
-import net.sf.l2j.gameserver.model.actor.instance.L2PetInstance;
+import net.sf.l2j.gameserver.model.actor.Playable;
+import net.sf.l2j.gameserver.model.actor.instance.Pet;
+import net.sf.l2j.gameserver.model.actor.instance.Player;
 import net.sf.l2j.gameserver.model.item.instance.ItemInstance;
 import net.sf.l2j.gameserver.network.SystemMessageId;
 import net.sf.l2j.gameserver.network.serverpackets.MagicSkillUse;
 import net.sf.l2j.gameserver.network.serverpackets.SystemMessage;
-import net.sf.l2j.gameserver.util.Util;
 
 /**
  * @author Kerberos
@@ -34,7 +18,7 @@ import net.sf.l2j.gameserver.util.Util;
 public class PetFood implements IItemHandler
 {
 	@Override
-	public void useItem(L2Playable playable, ItemInstance item, boolean forceUse)
+	public void useItem(Playable playable, ItemInstance item, boolean forceUse)
 	{
 		int itemId = item.getItemId();
 		switch (itemId)
@@ -60,14 +44,14 @@ public class PetFood implements IItemHandler
 		}
 	}
 	
-	public boolean useFood(L2Playable activeChar, int magicId, ItemInstance item)
+	public boolean useFood(Playable activeChar, int magicId, ItemInstance item)
 	{
 		L2Skill skill = SkillTable.getInstance().getInfo(magicId, 1);
 		if (skill != null)
 		{
-			if (activeChar instanceof L2PetInstance)
+			if (activeChar instanceof Pet)
 			{
-				L2PetInstance pet = (L2PetInstance) activeChar;
+				Pet pet = (Pet) activeChar;
 				if (pet.destroyItem("Consume", item.getObjectId(), 1, null, false))
 				{
 					// Send visual effect.
@@ -77,31 +61,25 @@ public class PetFood implements IItemHandler
 					pet.setCurrentFed(pet.getCurrentFed() + (skill.getFeed() * Config.PET_FOOD_RATE));
 					
 					// If pet is still hungry, send an alert.
-					if (pet.getCurrentFed() < (55 / 100f * pet.getPetLevelData().getPetMaxFeed()))
+					if (pet.checkAutoFeedState())
 						pet.getOwner().sendPacket(SystemMessageId.YOUR_PET_ATE_A_LITTLE_BUT_IS_STILL_HUNGRY);
 					
 					return true;
 				}
 			}
-			else if (activeChar instanceof L2PcInstance)
+			else if (activeChar instanceof Player)
 			{
-				L2PcInstance player = ((L2PcInstance) activeChar);
-				int itemId = item.getItemId();
-				if (player.isMounted())
+				final Player player = ((Player) activeChar);
+				final int itemId = item.getItemId();
+				
+				if (player.isMounted() && player.getPetTemplate().canEatFood(itemId))
 				{
-					int food[] = PetDataTable.getInstance().getPetData(player.getMountNpcId()).getFood();
-					if (Util.contains(food, itemId))
+					if (player.destroyItem("Consume", item.getObjectId(), 1, null, false))
 					{
-						if (player.destroyItem("Consume", item.getObjectId(), 1, null, false))
-						{
-							player.broadcastPacket(new MagicSkillUse(activeChar, activeChar, magicId, 1, 0, 0));
-							player.setCurrentFeed(player.getCurrentFeed() + (skill.getFeed() * Config.PET_FOOD_RATE));
-						}
-						return true;
+						player.broadcastPacket(new MagicSkillUse(activeChar, activeChar, magicId, 1, 0, 0));
+						player.setCurrentFeed(player.getCurrentFeed() + (skill.getFeed() * Config.PET_FOOD_RATE));
 					}
-					
-					activeChar.sendPacket(SystemMessage.getSystemMessage(SystemMessageId.S1_CANNOT_BE_USED).addItemName(itemId));
-					return false;
+					return true;
 				}
 				
 				activeChar.sendPacket(SystemMessage.getSystemMessage(SystemMessageId.S1_CANNOT_BE_USED).addItemName(itemId));
